// src/api/client.ts
export interface IOrganizationClient {

    /**
     * @return OK
     */
    organization_Get(signal?: AbortSignal): Promise<Organization[]>;

    /**
     * @return OK
     */
    organization_GetById(id: string, signal?: AbortSignal): Promise<Organization>;
}

export class OrganizationClient implements IOrganizationClient {
    private http: { fetch(url: RequestInfo, init?: RequestInit): Promise<Response> };
    private baseUrl: string;
    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;

    constructor(baseUrl?: string, http?: { fetch(url: RequestInfo, init?: RequestInit): Promise<Response> }) {
        this.http = http ? http : window as any;
        this.baseUrl = baseUrl ?? "";
    }

    /**
     * @return OK
     */
    organization_Get(signal?: AbortSignal): Promise<Organization[]> {
        let url_ = this.baseUrl + "/api/Organization";
        url_ = url_.replace(/[?&]$/, "");

        let options_: RequestInit = {
            method: "GET",
            signal,
            headers: {
                "Accept": "application/json"
            }
        };

        return this.http.fetch(url_, options_).then((_response: Response) => {
            return this.processOrganization_Get(_response);
        });
    }

    protected processOrganization_Get(response: Response): Promise<Organization[]> {
        const status = response.status;
        let _headers: any = {}; if (response.headers && response.headers.forEach) { response.headers.forEach((v: any, k: any) => _headers[k] = v); };
        if (status === 200) {
            return response.text().then((_responseText) => {
            let result200: any = null;
            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
            if (Array.isArray(resultData200)) {
                result200 = [] as any;
                for (let item of resultData200)
                    result200!.push(Organization.fromJS(item));
            }
            else {
                result200 = null as any;
            }
            return result200;
            });
        } else if (status === 400) {
            return response.text().then((_responseText) => {
            let result400: any = null;
            let resultData400 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
                result400 = resultData400 !== undefined ? resultData400 : null as any;
    
            return throwException("Bad Request", status, _responseText, _headers, result400);
            });
        } else if (status !== 200 && status !== 204) {
            return response.text().then((_responseText) => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            });
        }
        return Promise.resolve<Organization[]>(null as any);
    }

    /**
     * @return OK
     */
    organization_GetById(id: string, signal?: AbortSignal): Promise<Organization> {
        let url_ = this.baseUrl + "/api/Organization/{id}";
        if (id === undefined || id === null)
            throw new globalThis.Error("The parameter 'id' must be defined.");
        url_ = url_.replace("{id}", encodeURIComponent("" + id));
        url_ = url_.replace(/[?&]$/, "");

        let options_: RequestInit = {
            method: "GET",
            signal,
            headers: {
                "Accept": "application/json"
            }
        };

        return this.http.fetch(url_, options_).then((_response: Response) => {
            return this.processOrganization_GetById(_response);
        });
    }

    protected processOrganization_GetById(response: Response): Promise<Organization> {
        const status = response.status;
        let _headers: any = {}; if (response.headers && response.headers.forEach) { response.headers.forEach((v: any, k: any) => _headers[k] = v); };
        if (status === 200) {
            return response.text().then((_responseText) => {
            let result200: any = null;
            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
            result200 = Organization.fromJS(resultData200);
            return result200;
            });
        } else if (status === 404) {
            return response.text().then((_responseText) => {
            let result404: any = null;
            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
                result404 = resultData404 !== undefined ? resultData404 : null as any;
    
            return throwException("Not Found", status, _responseText, _headers, result404);
            });
        } else if (status !== 200 && status !== 204) {
            return response.text().then((_responseText) => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            });
        }
        return Promise.resolve<Organization>(null as any);
    }
}
// ...
export class AdditionalTask implements IAdditionalTask {
    businessLine!: BusinessLine;
    organization!: string;
    taskDescription!: string;
    projectRequestId!: number;

    constructor(data?: IAdditionalTask) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.businessLine = _data["businessLine"];
            this.organization = _data["organization"];
            this.taskDescription = _data["taskDescription"];
            this.projectRequestId = _data["projectRequestId"];
        }
    }

    static fromJS(data: any): AdditionalTask {
        data = typeof data === 'object' ? data : {};
        let result = new AdditionalTask();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["businessLine"] = this.businessLine;
        data["organization"] = this.organization;
        data["taskDescription"] = this.taskDescription;
        data["projectRequestId"] = this.projectRequestId;
        return data;
    }
}

export interface IAdditionalTask {
    businessLine: BusinessLine;
    organization: string;
    taskDescription: string;
    projectRequestId: number;
}

export class StageData implements IStageData {

    constructor(data?: IStageData) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
    }

    static fromJS(data: any): StageData {
        data = typeof data === 'object' ? data : {};
        let result = new StageData();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        return data;
    }
}

export interface IStageData {
}
// ...
export class ClientDto implements IClientDto {
    clientId!: string;
    name!: string;

    constructor(data?: IClientDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.clientId = _data["clientId"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): ClientDto {
        data = typeof data === 'object' ? data : {};
        let result = new ClientDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["clientId"] = this.clientId;
        data["name"] = this.name;
        return data;
    }
}

export interface IClientDto {
    clientId: string;
    name: string;
}

export type ContractTypeCategory = "Master" | "Stand Alone" | "Task Order" | "Cost/Fee Proposal" | "Undetermined" | "Historical Data";

export class EmployeeDto implements IEmployeeDto {
    employee!: string;
    titleName!: string;

    constructor(data?: IEmployeeDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.employee = _data["employee"];
            this.titleName = _data["titleName"];
        }
    }

    static fromJS(data: any): EmployeeDto {
        data = typeof data === 'object' ? data : {};
        let result = new EmployeeDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["employee"] = this.employee;
        data["titleName"] = this.titleName;
        return data;
    }
}

export interface IEmployeeDto {
    employee: string;
    titleName: string;
}
// ...
export class Organization implements IOrganization {
    org!: string;
    name!: string;

    constructor(data?: IOrganization) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.org = _data["org"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): Organization {
        data = typeof data === 'object' ? data : {};
        let result = new Organization();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["org"] = this.org;
        data["name"] = this.name;
        return data;
    }
}

export interface IOrganization {
    org: string;
    name: string;
}
// ...
export class ProjectRequest extends ProjectRequestBase implements IProjectRequest {
    additionalTasks!: AdditionalTask[];
    stageData!: StageData;

    constructor(data?: IProjectRequest) {
        super(data);
        if (!data) {
            this.additionalTasks = [];
            this.stageData = new StageData();
        }
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            if (Array.isArray(_data["additionalTasks"])) {
                this.additionalTasks = [] as any;
                for (let item of _data["additionalTasks"])
                    this.additionalTasks!.push(AdditionalTask.fromJS(item));
            }
            this.stageData = _data["stageData"] ? StageData.fromJS(_data["stageData"]) : new StageData();
        }
    }

    static override fromJS(data: any): ProjectRequest {
        data = typeof data === 'object' ? data : {};
        let result = new ProjectRequest();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.additionalTasks)) {
            data["additionalTasks"] = [];
            for (let item of this.additionalTasks)
                data["additionalTasks"].push(item ? item.toJSON() : undefined as any);
        }
        data["stageData"] = this.stageData ? this.stageData.toJSON() : undefined as any;
        super.toJSON(data);
        return data;
    }
}

export interface IProjectRequest extends IProjectRequestBase {
    additionalTasks: AdditionalTask[];
    stageData: StageData;
}
// ...
export type ProjectRequestStage = "Pending" | "Approved" | "Changes Requested";